+++ codeceptConfigHelper.js
@@ -11,6 +11,7 @@ module.exports = {
     alertsPage: './tests/ia/pages/alertsPage.js',
     allChecksPage: './tests/stt/pages/allChecksPage.js',
     amiInstanceSetupPage: './tests/pages/amiInstanceSetupPage.js',
+    annotationAPI: './tests/pages/api/annotationAPI.js',
     channelsAPI: './tests/ia/pages/api/channelsAPI.js',
     dashboardPage: './tests/pages/dashboardPage.js',
     databaseChecksPage: './tests/stt/pages/databaseChecksPage.js',
